move component to separate file
